folders/files views {-||-} early impl...
