+++ administrator/components/com_media/views/files/tmpl/default.php
@@ -32,11 +32,6 @@ JFactory::getDocument()->addStyleDeclaration(
 ?>
 <?php if (count($this->images) > 0 || count($this->folders) > 0) : ?>
 	<ul class="manager thumbnails">
-		<?php for ($i = 0, $n = count($this->folders); $i < $n; $i++) :
-			$this->setFolder($i);
-			echo $this->loadTemplate('folder');
-		endfor; ?>
-
 		<?php for ($i = 0, $n = count($this->images); $i < $n; $i++) :
 			$this->setImage($i);
 			echo $this->loadTemplate('image');
