+++ administrator/components/com_media/views/files/tmpl/default_image.php
@@ -17,7 +17,7 @@ $dispatcher->trigger('onContentBeforeDisplay', array('com_media.file', &$this->_
 ?>
 
 <li class="imgOutline thumbnail height-80 width-80 center">
-	<a class="img-preview" href="javascript:ImageManager.populateFields('<?php echo $this->_tmp_img->path_relative; ?>')" title="<?php echo $this->_tmp_img->name; ?>" >
+	<a class="img-preview" href="javascript:ImageManager.populateFields('<?php //echo $this->_tmp_img->path_relative; ?>')" title="<?php //echo $this->_tmp_img->name; ?>" >
 		<div class="height-50">
 			<?php echo JHtml::_('image', $this->baseURL . '/' . $this->_tmp_img->path_relative, JText::sprintf('COM_MEDIA_IMAGE_TITLE', $this->_tmp_img->title, JHtml::_('number.bytes', $this->_tmp_img->size)), array('width' => $this->_tmp_img->width_60, 'height' => $this->_tmp_img->height_60)); ?>
 		</div>
