--- administrator/components/com_media/views/images/view.html.php
@@ -1,46 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_media
- *
- * @copyright   Copyright (C) 2005 - 2015 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-/**
- * HTML View class for the Media component
- *
- * @since  1.0
- */
-class MediaViewImages extends JViewLegacy
-{
-	/**
-	 * Execute and display a template script.
-	 *
-	 * @param   string  $tpl  The name of the template file to parse; automatically searches through the template paths.
-	 *
-	 * @return  mixed  A string if successful, otherwise a Error object.
-	 *
-	 * @since   1.0
-	 */
-	public function display($tpl = null)
-	{
-		$config = JComponentHelper::getParams('com_media');
-
-		/*
-		 * Display form for FTP credentials?
-		 * Don't set them here, as there are other functions called before this one if there is any file write operation
-		 */
-		$ftp = !JClientHelper::hasCredentials('ftp');
-
-		$this->session     = JFactory::getSession();
-		$this->config      = $config;
-		$this->state       = $this->get('state');
-		$this->folderList  = $this->get('folderList');
-		$this->require_ftp = $ftp;
-
-		parent::display($tpl);
-	}
-}
