--- administrator/components/com_media/views/imageslist/tmpl/default_folder.php
@@ -1,23 +1,0 @@
-<?php
-/**
- * @package     Joomla.Administrator
- * @subpackage  com_media
- *
- * @copyright   Copyright (C) 2005 - 2015 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-defined('_JEXEC') or die;
-
-$input = JFactory::getApplication()->input;
-?>
-<li class="imgOutline thumbnail height-80 width-80 center">
-	<a href="index.php?option=com_media&amp;view=imagesList&amp;tmpl=component&amp;folder=<?php echo $this->_tmp_folder->path_relative; ?>&amp;asset=<?php echo $input->getCmd('asset');?>&amp;author=<?php echo $input->getCmd('author');?>" target="imageframe">
-		<div class="height-50">
-			<span class="icon-folder-2"></span>
-		</div>
-		<div class="small">
-			<?php echo JHtml::_('string.truncate', $this->_tmp_folder->name, 10, false); ?>
-		</div>
-	</a>
-</li>
